+++ src/lib/blog/fetchAllPosts/splitDirAndFiles.ts
